Refactoring SwipeCard component/container structure
