+++ FoodSwipe/containers/SwipeCards.js
@@ -2,25 +2,10 @@
 'use strict';
 
 import React, { Component } from 'react';
-import {StyleSheet, Text, View, Image} from 'react-native';
 
 import SwipeCards from 'react-native-swipe-cards';
 import NoMoreCards from '../components/NoMoreCards'
-// import RestaurantCard from '../components/RestrauntCard';
-
-class Card extends Component {
-  constructor(props) {
-    super(props);
-  }
-	//TODO: Use restaurantProfile for card
-  render() {
-    return (
-      <View style={[styles.card, {backgroundColor: this.props.backgroundColor}]}>
-        <Text>{this.props.text}</Text>
-      </View>
-    )
-  }
-}
+import RestaurantCard from '../components/RestaurantCard';
 
 export default class extends React.Component {
   constructor(props) {
@@ -49,26 +34,13 @@ export default class extends React.Component {
     return (
       <SwipeCards
         cards={this.state.cards}
-        renderCard={(cardData) => <Card {...cardData} />}
+        renderCard={(cardData) => <RestaurantCard {...cardData} />}
         renderNoMoreCards={() => <NoMoreCards />}
 
 				yupText={"YUM!"}
         handleYup={this.handleYup}
         handleNope={this.handleNope}
-
       />
     )
   }
 }
-
-const styles = StyleSheet.create({
-  card: {
-    justifyContent: 'center',
-    alignItems: 'center',
-		flex: 1,
-		minWidth: 350,
-		marginTop: 75,
-		marginHorizontal: 25,
-		marginBottom: 50,
-  },
-})
